[BPK-4385]: Migrate description list and dialog
